uzhavar-neradi/backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-otp==1.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
kombu==5.6.2
packaging==26.0
phonenumberslite==9.0.26
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
redis==7.3.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-otp==1.3.0
phonenumberslite==8.13.27
Pillow==10.1.0
django-environ==0.11.2
gunicorn==21.2.0